added some new apps to future development
