+++ apps/__init__.py
@@ -1,3 +1,3 @@
-from apps.order import *
-from apps.product import *
+from apps.orders import *
+from apps.products import *
 from apps.user_profile import *
