+++ apps/orders/apps.py
@@ -1,6 +1,6 @@
 from django.apps import AppConfig
 
 
-class ProductConfig(AppConfig):
+class OrderConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'product'
+    name = 'apps.orders'
