+++ apps/orders/models/__init__.py
@@ -1,0 +1,2 @@
+from .order import Order
+from .order_itens import OrderItem
