+++ apps/products/admin.py
